+++ mpicbg/imglib/container/dynamic/ByteDynamicContainer.java
@@ -39,7 +39,7 @@ public class ByteDynamicContainer <T extends Type<T>> extends DynamicContainer<T
 {
 	final ArrayList<Byte> data;
 	
-	public ByteDynamicContainer( final DynamicContainerFactory factory, final long[] dim, final int entitiesPerPixel )
+	public ByteDynamicContainer( final DynamicContainerFactory factory, final int[] dim, final int entitiesPerPixel )
 	{
 		super( factory, dim, entitiesPerPixel );
