+++ mpicbg/imglib/container/dynamic/LongDynamicContainer.java
@@ -39,7 +39,7 @@ public class LongDynamicContainer <T extends Type<T>> extends DynamicContainer<T
 {
 	final ArrayList<Long> data;
 	
-	public LongDynamicContainer( final DynamicContainerFactory factory, final long[] dim, final int entitiesPerPixel )
+	public LongDynamicContainer( final DynamicContainerFactory factory, final int[] dim, final int entitiesPerPixel )
 	{
 		super( factory, dim, entitiesPerPixel );
